Fix description field not being omitted on issue creation
This happens because, if the description is omitted by the user, the Body field in jira.CreateRequest
is set to the zero value of string when the CreateRequest struct is created. But the Body field
is an interface{} instead of string (to handle ADF.adf polymorphism) so the omitempty option from json package doesn't work, because the considered zero value in this case is nil.

To solve this, I introduced an helper function (cr*).ForBody(body string) that will set the Body field to
nil when the body string is a blank string instead of the empty string.

The result is that the description field is properly omitted in the JSON request when the user
gives no input for the description.